dailybuild: warningの抑止

+++ core/src/HumanInterface/HISpidar.cpp
@@ -151,7 +151,7 @@ void HISpidar4Desc::InitSpidarG(char* type){
 		for (int i = 0; i < motors.size(); ++i) {
 			motors[i].currentPerVolt = 20.0f;	//	1A / 0.05 OHM
 			motors[i].voltPerNewton = (1.0f / motors[i].currentPerVolt) * (1.0f / 10.0f);	//	V/N = V/C * C/N
-			motors[i].lengthPerPulse = 0.004f * M_PI / 1024.0f;	 //	D=4mm * PI / (count/revolution)
+			motors[i].lengthPerPulse = 0.004f * (float) M_PI / 1024.0f;	 //	D=4mm * PI / (count/revolution)
 		}
 		nButton = 0;
 	}
